requirements.txt
fastapi
uvicorn
python-telegram-bot
requests
python-dotenv
google-generativeai
yfinance
pymongo